# [mypyc] Implement dict copy primitive