@vsakkas 2020-11-14
It seems a test case (`testTypedDictClassInheritanceWithTotalArgument`), which is (seemingly) unrelated to this PR is failing:
```
Expected:
  main:9: note: Revealed type is 'TypedDict('__main__.C', {'x': builtins.int, 'y'?: builtins.int, 'z': builtins.str})' (diff)
Actual:
  main:9: note: Revealed type is 'TypedDict('__main__.C', {'y'?: builtins.int, 'x': builtins.int, 'z': builtins.str})' (diff)
```
Any pointers as to why this could be happening?