@vsakkas 2020-11-18
Results from https://github.com/mypyc/mypyc-benchmarks:

Master:
```
running dict_copy
..........
interpreted: 0.296086s (avg of 5 iterations; stdev 1.1%)
compiled:    0.185336s (avg of 5 iterations; stdev 0.67%)

compiled is 1.598x faster
```

PR:
```
running dict_copy
..........
interpreted: 0.300922s (avg of 5 iterations; stdev 0.42%)
compiled:    0.131764s (avg of 5 iterations; stdev 0.91%)

compiled is 2.284x faster
```